Move leaf nodes to separate file
